Refactor wasmtime and wazero into Runners
